chore(tests): port and re-enable `TestAdmissionWebhook_KongConsumers` and deduplicate code
