+++ ingress-controller/test/envtest/admission_webhook_envtest_test.go
@@ -2,17 +2,14 @@ package envtest
 
 import (
 	"context"
-	"errors"
 	"testing"
 	"time"
 
-	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	admregv1 "k8s.io/api/admissionregistration/v1"
 	corev1 "k8s.io/api/core/v1"
 	apiextensionsv1 "k8s.io/apiextensions-apiserver/pkg/apis/apiextensions/v1"
-	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	k8stypes "k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -526,337 +523,6 @@ func TestAdmissionWebhook_KongClusterPlugins(t *testing.T) {
 				}
 			}, waitTime, tickTime,
 			)
-		})
-	}
-}
-
-func TestAdmissionWebhook_KongConsumers(t *testing.T) {
-	t.Skip("skipping until https://github.com/Kong/kong-operator/issues/2176 is resolved")
-	ctx, cancel := context.WithCancel(t.Context())
-	defer cancel()
-
-	var (
-		scheme           = Scheme(t, WithKong)
-		envcfg           = Setup(t, scheme)
-		ctrlClientGlobal = NewControllerClient(t, scheme, envcfg)
-		ns               = CreateNamespace(ctx, t, ctrlClientGlobal)
-		ctrlClient       = client.NewNamespacedClient(ctrlClientGlobal, ns.Name)
-
-		kongContainer = runKongEnterprise(ctx, t)
-	)
-
-	logs := RunManager(ctx, t, envcfg,
-		AdminAPIOptFns(),
-		WithPublishService(ns.Name),
-		WithKongAdminURLs(kongContainer.AdminURL(ctx, t)),
-	)
-	WaitForManagerStart(t, logs)
-
-	t.Logf("creating some static credentials in %s namespace which will be used to test global validation", ns.Name)
-	for _, secret := range []*corev1.Secret{
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "tuxcreds1",
-				Labels: map[string]string{
-					labels.CredentialTypeLabel: "basic-auth",
-				},
-			},
-			StringData: map[string]string{
-				"username": "tux1",
-				"password": "testpass",
-			},
-		},
-		{
-			ObjectMeta: metav1.ObjectMeta{
-				Name: "tuxcreds2",
-				Labels: map[string]string{
-					labels.CredentialTypeLabel: "basic-auth",
-				},
-			},
-			StringData: map[string]string{
-				"username": "tux2",
-				"password": "testpass",
-			},
-		},
-	} {
-		secret := secret.DeepCopy()
-		require.NoError(t, ctrlClient.Create(ctx, secret))
-		t.Cleanup(func() {
-			if err := ctrlClient.Delete(ctx, secret); err != nil && !apierrors.IsNotFound(err) && !errors.Is(err, context.Canceled) {
-				assert.NoError(t, err)
-			}
-		})
-	}
-
-	t.Logf("creating a static consumer in %s namespace which will be used to test global validation", ns.Name)
-	consumer := &configurationv1.KongConsumer{
-		ObjectMeta: metav1.ObjectMeta{
-			GenerateName: "statis-consumer-",
-			Annotations: map[string]string{
-				annotations.IngressClassKey: annotations.DefaultIngressClass,
-			},
-		},
-		Username: "tux",
-		CustomID: uuid.NewString(),
-		Credentials: []string{
-			"tuxcreds1",
-			"tuxcreds2",
-		},
-	}
-	require.NoError(t, ctrlClient.Create(ctx, consumer))
-	t.Cleanup(func() {
-		ctx := t.Context()
-		if err := ctrlClient.Delete(ctx, consumer); err != nil && !apierrors.IsNotFound(err) && !errors.Is(err, context.Canceled) {
-			assert.NoError(t, err)
-		}
-	})
-
-	testCases := []struct {
-		name           string
-		consumer       *configurationv1.KongConsumer
-		credentials    []*corev1.Secret
-		wantErr        bool
-		wantPartialErr string
-	}{
-		{
-			name: "a consumer with no credentials should pass validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "testconsumer",
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username: uuid.NewString(),
-				CustomID: uuid.NewString(),
-			},
-			credentials: nil,
-			wantErr:     false,
-		},
-		{
-			name: "a consumer with valid credentials should pass validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: uuid.NewString(),
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username:    "electron",
-				CustomID:    uuid.NewString(),
-				Credentials: []string{"electronscreds"},
-			},
-			credentials: []*corev1.Secret{{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "electronscreds",
-					Labels: map[string]string{
-						labels.CredentialTypeLabel: "basic-auth",
-					},
-				},
-				StringData: map[string]string{
-					"username": "electron",
-					"password": "testpass",
-				},
-			}},
-			wantErr: false,
-		},
-		{
-			name: "a consumer with duplicate credentials which are NOT constrained should pass validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: uuid.NewString(),
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username: "proton",
-				CustomID: uuid.NewString(),
-				Credentials: []string{
-					"protonscreds1",
-					"protonscreds2",
-				},
-			},
-			credentials: []*corev1.Secret{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "protonscreds1",
-						Labels: map[string]string{
-							labels.CredentialTypeLabel: "basic-auth",
-						},
-					},
-					StringData: map[string]string{
-						"username": "proton",
-						"password": "testpass",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "protonscreds2",
-						Labels: map[string]string{
-							labels.CredentialTypeLabel: "basic-auth",
-						},
-					},
-					StringData: map[string]string{
-						"username": "electron", // username is unique constrained
-						"password": "testpass", // password is not unique constrained
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "a consumer referencing credentials secrets which do not yet exist should fail validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: uuid.NewString(),
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username: "repairedlawnmower",
-				CustomID: uuid.NewString(),
-				Credentials: []string{
-					"nonexistentcreds",
-				},
-			},
-			wantErr:        true,
-			wantPartialErr: "not found",
-		},
-		{
-			name: "a consumer with duplicate credentials which ARE constrained should fail validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "brokenshovel",
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username: "neutron",
-				CustomID: uuid.NewString(),
-				Credentials: []string{
-					"neutronscreds1",
-					"neutronscreds2",
-				},
-			},
-			credentials: []*corev1.Secret{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "neutronscreds1",
-						Labels: map[string]string{
-							labels.CredentialTypeLabel: "basic-auth",
-						},
-					},
-					StringData: map[string]string{
-						"username": "neutron",
-						"password": "testpass",
-					},
-				},
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "neutronscreds2",
-						Labels: map[string]string{
-							labels.CredentialTypeLabel: "basic-auth",
-						},
-					},
-					StringData: map[string]string{
-						"username": "neutron", // username is unique constrained
-						"password": "testpass",
-					},
-				},
-			},
-			wantErr:        true,
-			wantPartialErr: "unique key constraint violated for username",
-		},
-		{
-			name: "a consumer that provides duplicate credentials which are NOT in violation of unique key constraints should pass validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: uuid.NewString(),
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username: "reasonablehammer",
-				CustomID: uuid.NewString(),
-				Credentials: []string{
-					"reasonablehammer",
-				},
-			},
-			credentials: []*corev1.Secret{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "reasonablehammer",
-						Labels: map[string]string{
-							labels.CredentialTypeLabel: "basic-auth",
-						},
-					},
-					StringData: map[string]string{
-						"username": "reasonablehammer",
-						"password": "testpass", // not unique constrained, so even though someone else is using this password this should pass
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name: "a consumer that provides credentials that are in violation of unique constraints globally against other existing consumers should fail validation",
-			consumer: &configurationv1.KongConsumer{
-				ObjectMeta: metav1.ObjectMeta{
-					GenerateName: "violating-uniqueness-",
-					Annotations: map[string]string{
-						annotations.IngressClassKey: annotations.DefaultIngressClass,
-					},
-				},
-				Username: "unreasonablehammer",
-				CustomID: uuid.NewString(),
-				Credentials: []string{
-					"unreasonablehammer",
-				},
-			},
-			credentials: []*corev1.Secret{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "unreasonablehammer",
-						Labels: map[string]string{
-							labels.CredentialTypeLabel: "basic-auth",
-						},
-					},
-					StringData: map[string]string{
-						"username": "tux1", // unique constrained with previous created static consumer credentials
-						"password": "testpass",
-					},
-				},
-			},
-			wantErr:        true,
-			wantPartialErr: "unique key constraint violated for username",
-		},
-	}
-	for _, tc := range testCases {
-		t.Run(tc.name, func(t *testing.T) {
-			for _, credential := range tc.credentials {
-				require.NoError(t, ctrlClient.Create(ctx, credential))
-				t.Cleanup(func() { //nolint:contextcheck
-					ctx := context.Background()
-					if err := ctrlClient.Delete(ctx, credential); err != nil && !apierrors.IsNotFound(err) {
-						assert.NoError(t, err)
-					}
-				})
-			}
-
-			err := ctrlClient.Create(ctx, tc.consumer)
-			if tc.wantErr {
-				require.Error(t, err, "consumer %s should fail to create", tc.consumer.Name)
-				assert.Contains(t, err.Error(), tc.wantPartialErr,
-					"got error string %q, want a superstring of %q", err.Error(), tc.wantPartialErr,
-				)
-			} else {
-				t.Cleanup(func() {
-					if err := ctrlClient.Delete(ctx, tc.consumer); err != nil && !apierrors.IsNotFound(err) {
-						assert.NoError(t, err)
-					}
-				})
-				require.NoError(t, err, "consumer %s should create successfully", tc.consumer.Name)
-			}
 		})
 	}
 }
